Added test to reveal next problem with records: specific constructor/accessor in stubs that should override the record component.
