+++ checker/tests/stubparser-records/RecordStubbed.astub
@@ -1,0 +1,14 @@
+import org.checkerframework.checker.nullness.qual.NonNull;
+import org.checkerframework.checker.nullness.qual.Nullable;
+
+/**
+ * nxx is nullable in the record but non-null in constructor and accessor via stubs
+ * nsxx is nullable in the stubs but non-null in constructor and accessor via stubs
+ * xnn is de-facto non-null in record but nullable in constructor and accessor via stubs
+ */
+public record RecordStubbed(String nxx, @Nullable String nsxx, Integer xnn) {
+  RecordStubbed(@NonNull String nxx, @NonNull String nsxx, @Nullable Integer xnn);
+  @NonNull String nxx();
+  @NonNull String nsxx();
+  @Nullable Integer xnn();
+}
